backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
mcp
httpx
email-validator
pyjwt
cryptography
